Add tests for meshlib add command

+++ tests/temp_dir.py
@@ -5,9 +5,12 @@ from pathlib import Path
 
 
 @contextmanager
-def TEMP_DIR(root: Path = Path("."), delete: bool = True):
+def TEMP_DIR(
+    root: Path = Path("."), delete: bool = True, pre_delete: bool = True
+):
     temp_dir = root / ".temp"
-    shutil.rmtree(temp_dir, True)
+    if pre_delete:
+        shutil.rmtree(temp_dir, True)
     temp_dir.mkdir(0o777, True, True)
     yield temp_dir
     if delete:
